* develop:
  Adjusted colors on engine, JJ and heatsink.
  Adjusted UI colors
  Solving issue: #216.
  Formatting code, dotting i-s and crossing t-s.
  Changed colors for bar, dynamic, energy_alt.
  Updated ignore file and added more styling options.

+++ src/lisong_mechlab/model/mwo_parsing/helpers/ItemStatsModuleStats.java
@@ -27,5 +27,5 @@ public class ItemStatsModuleStats{
    @XStreamAsAttribute
    public double tons;
    @XStreamAsAttribute
-   public double health;
+   public int    health;
 }
